+++ packages/atproto/lib/src/entities/parsed_service.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'service.dart';
+part of 'parsed_service.dart';
 
 // **************************************************************************
 // FreezedGenerator
@@ -14,12 +14,12 @@ T _$identity<T>(T value) => value;
 final _privateConstructorUsedError = UnsupportedError(
     'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
 
-Service _$ServiceFromJson(Map<String, dynamic> json) {
-  return _Service.fromJson(json);
+ParsedService _$ParsedServiceFromJson(Map<String, dynamic> json) {
+  return _ParsedService.fromJson(json);
 }
 
 /// @nodoc
-mixin _$Service {
+mixin _$ParsedService {
   String get id => throw _privateConstructorUsedError;
   String get type => throw _privateConstructorUsedError;
   @JsonKey(name: 'serviceEndpoint')
@@ -27,13 +27,15 @@ mixin _$Service {
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
-  $ServiceCopyWith<Service> get copyWith => throw _privateConstructorUsedError;
+  $ParsedServiceCopyWith<ParsedService> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
-abstract class $ServiceCopyWith<$Res> {
-  factory $ServiceCopyWith(Service value, $Res Function(Service) then) =
-      _$ServiceCopyWithImpl<$Res, Service>;
+abstract class $ParsedServiceCopyWith<$Res> {
+  factory $ParsedServiceCopyWith(
+          ParsedService value, $Res Function(ParsedService) then) =
+      _$ParsedServiceCopyWithImpl<$Res, ParsedService>;
   @useResult
   $Res call(
       {String id,
@@ -42,9 +44,9 @@ abstract class $ServiceCopyWith<$Res> {
 }
 
 /// @nodoc
-class _$ServiceCopyWithImpl<$Res, $Val extends Service>
-    implements $ServiceCopyWith<$Res> {
-  _$ServiceCopyWithImpl(this._value, this._then);
+class _$ParsedServiceCopyWithImpl<$Res, $Val extends ParsedService>
+    implements $ParsedServiceCopyWith<$Res> {
+  _$ParsedServiceCopyWithImpl(this._value, this._then);
 
   // ignore: unused_field
   final $Val _value;
@@ -76,10 +78,11 @@ class _$ServiceCopyWithImpl<$Res, $Val extends Service>
 }
 
 /// @nodoc
-abstract class _$$_ServiceCopyWith<$Res> implements $ServiceCopyWith<$Res> {
-  factory _$$_ServiceCopyWith(
-          _$_Service value, $Res Function(_$_Service) then) =
-      __$$_ServiceCopyWithImpl<$Res>;
+abstract class _$$_ParsedServiceCopyWith<$Res>
+    implements $ParsedServiceCopyWith<$Res> {
+  factory _$$_ParsedServiceCopyWith(
+          _$_ParsedService value, $Res Function(_$_ParsedService) then) =
+      __$$_ParsedServiceCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call(
@@ -89,10 +92,11 @@ abstract class _$$_ServiceCopyWith<$Res> implements $ServiceCopyWith<$Res> {
 }
 
 /// @nodoc
-class __$$_ServiceCopyWithImpl<$Res>
-    extends _$ServiceCopyWithImpl<$Res, _$_Service>
-    implements _$$_ServiceCopyWith<$Res> {
-  __$$_ServiceCopyWithImpl(_$_Service _value, $Res Function(_$_Service) _then)
+class __$$_ParsedServiceCopyWithImpl<$Res>
+    extends _$ParsedServiceCopyWithImpl<$Res, _$_ParsedService>
+    implements _$$_ParsedServiceCopyWith<$Res> {
+  __$$_ParsedServiceCopyWithImpl(
+      _$_ParsedService _value, $Res Function(_$_ParsedService) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -102,7 +106,7 @@ class __$$_ServiceCopyWithImpl<$Res>
     Object? type = null,
     Object? endpoint = null,
   }) {
-    return _then(_$_Service(
+    return _then(_$_ParsedService(
       id: null == id
           ? _value.id
           : id // ignore: cast_nullable_to_non_nullable
@@ -121,14 +125,14 @@ class __$$_ServiceCopyWithImpl<$Res>
 
 /// @nodoc
 @JsonSerializable()
-class _$_Service implements _Service {
-  const _$_Service(
+class _$_ParsedService implements _ParsedService {
+  const _$_ParsedService(
       {required this.id,
       required this.type,
       @JsonKey(name: 'serviceEndpoint') required this.endpoint});
 
-  factory _$_Service.fromJson(Map<String, dynamic> json) =>
-      _$$_ServiceFromJson(json);
+  factory _$_ParsedService.fromJson(Map<String, dynamic> json) =>
+      _$$_ParsedServiceFromJson(json);
 
   @override
   final String id;
@@ -140,14 +144,14 @@ class _$_Service implements _Service {
 
   @override
   String toString() {
-    return 'Service(id: $id, type: $type, endpoint: $endpoint)';
+    return 'ParsedService(id: $id, type: $type, endpoint: $endpoint)';
   }
 
   @override
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_Service &&
+            other is _$_ParsedService &&
             (identical(other.id, id) || other.id == id) &&
             (identical(other.type, type) || other.type == type) &&
             (identical(other.endpoint, endpoint) ||
@@ -161,25 +165,26 @@ class _$_Service implements _Service {
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$_ServiceCopyWith<_$_Service> get copyWith =>
-      __$$_ServiceCopyWithImpl<_$_Service>(this, _$identity);
+  _$$_ParsedServiceCopyWith<_$_ParsedService> get copyWith =>
+      __$$_ParsedServiceCopyWithImpl<_$_ParsedService>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$_ServiceToJson(
+    return _$$_ParsedServiceToJson(
       this,
     );
   }
 }
 
-abstract class _Service implements Service {
-  const factory _Service(
+abstract class _ParsedService implements ParsedService {
+  const factory _ParsedService(
           {required final String id,
           required final String type,
           @JsonKey(name: 'serviceEndpoint') required final String endpoint}) =
-      _$_Service;
+      _$_ParsedService;
 
-  factory _Service.fromJson(Map<String, dynamic> json) = _$_Service.fromJson;
+  factory _ParsedService.fromJson(Map<String, dynamic> json) =
+      _$_ParsedService.fromJson;
 
   @override
   String get id;
@@ -190,6 +195,6 @@ abstract class _Service implements Service {
   String get endpoint;
   @override
   @JsonKey(ignore: true)
-  _$$_ServiceCopyWith<_$_Service> get copyWith =>
+  _$$_ParsedServiceCopyWith<_$_ParsedService> get copyWith =>
       throw _privateConstructorUsedError;
 }