+++ packages/atproto/lib/src/entities/parsed_service.g.dart
@@ -2,33 +2,29 @@
 
 // ignore_for_file: non_constant_identifier_names
 
-part of 'verification_method.dart';
+part of 'parsed_service.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_VerificationMethod _$$_VerificationMethodFromJson(Map json) =>
-    $checkedCreate(
-      r'_$_VerificationMethod',
+_$_ParsedService _$$_ParsedServiceFromJson(Map json) => $checkedCreate(
+      r'_$_ParsedService',
       json,
       ($checkedConvert) {
-        final val = _$_VerificationMethod(
+        final val = _$_ParsedService(
           id: $checkedConvert('id', (v) => v as String),
           type: $checkedConvert('type', (v) => v as String),
-          controller: $checkedConvert('controller', (v) => v as String),
-          publicKeyMultibase:
-              $checkedConvert('publicKeyMultibase', (v) => v as String),
+          endpoint: $checkedConvert('serviceEndpoint', (v) => v as String),
         );
         return val;
       },
+      fieldKeyMap: const {'endpoint': 'serviceEndpoint'},
     );
 
-Map<String, dynamic> _$$_VerificationMethodToJson(
-        _$_VerificationMethod instance) =>
+Map<String, dynamic> _$$_ParsedServiceToJson(_$_ParsedService instance) =>
     <String, dynamic>{
       'id': instance.id,
       'type': instance.type,
-      'controller': instance.controller,
-      'publicKeyMultibase': instance.publicKeyMultibase,
+      'serviceEndpoint': instance.endpoint,
     };